fix: More defensive code when trying to drop tables

+++ migration/2_sesha_upgrade_autoincrement.php
@@ -20,24 +20,20 @@ class SeshaUpgradeAutoIncrement extends Horde_Db_Migration_Base
     public function up()
     {
         $this->changeColumn('sesha_inventory', 'stock_id', 'autoincrementKey');
-        try {
+        if (in_array('sesha_inventory_seq', $this->tables())) {
             $this->dropTable('sesha_inventory_seq');
-        } catch (Horde_Db_Exception $e) {
         }
         $this->changeColumn('sesha_categories', 'category_id', 'autoincrementKey');
-        try {
+        if (in_array('sesha_categories_seq', $this->tables())) {
             $this->dropTable('sesha_categories_seq');
-        } catch (Horde_Db_Exception $e) {
         }
         $this->changeColumn('sesha_properties', 'property_id', 'autoincrementKey');
-        try {
+        if (in_array('sesha_properties_seq', $this->tables())) {
             $this->dropTable('sesha_properties_seq');
-        } catch (Horde_Db_Exception $e) {
         }
         $this->changeColumn('sesha_inventory_properties', 'attribute_id', 'autoincrementKey');
-        try {
+        if (in_array('sesha_inventory_properties_seq', $this->tables())) {
             $this->dropTable('sesha_inventory_properties_seq');
-        } catch (Horde_Db_Exception $e) {
         }
     }
